+++ PuxApp/Services/DirectoryScanner.cs
@@ -7,6 +7,12 @@ namespace PuxApp.Services
     /// </summary>
     public class DirectoryScanner
     {
+        private readonly DirectoryPathValidator _directoryPathValidator;
+        public DirectoryScanner(DirectoryPathValidator directoryPathValidator)
+        {
+            _directoryPathValidator = directoryPathValidator;
+        }
+
         /// <summary>
         /// Načte soubory z daného adresáře.
         /// </summary>
@@ -16,7 +22,7 @@ namespace PuxApp.Services
         /// <exception cref="DirectoryNotFoundException">Zadaná cesta je neplatná.</exception>
         public List<FileItem> ReturnFilesFromDirectory(string path)
         {
-            ValidatePath(path);
+            _directoryPathValidator.ValidatePath(path);
 
             return Directory
                 .GetFiles(path)
@@ -26,21 +32,6 @@ namespace PuxApp.Services
                     FullPath = file
                 })
                 .ToList();
-        }
-
-        /// <summary>
-        /// Ověří, zda je zadaná cesta platná a zda adresář existuje.
-        /// </summary>
-        /// <param name="path">Cesta k adresáři.</param>
-        /// <exception cref="ArgumentException">Zadaná cesta nesmí být prázdná.</exception>
-        /// <exception cref="DirectoryNotFoundException">Zadaná cesta je neplatná.</exception>
-        private static void ValidatePath(string path)
-        {
-            if (string.IsNullOrWhiteSpace(path))
-                throw new ArgumentException("Cesta nesmí být prázdná.");
-
-            if (!Directory.Exists(path))
-                throw new DirectoryNotFoundException("Zadaný adresář neexistuje.");
         }
     }
 }